[EI-231] [api] refactor: 폴더 구조 및 불필요 파일 refacotoring
