+++ api/src/numerical-guidance/api/history-indicator/history-indicator.controller.ts
@@ -5,9 +5,9 @@ import { ApiPaginatedResponseDecorator } from '../../../utils/pagination/api-pag
 import { ApiExceptionResponse } from '../../../utils/exception-filter/api-exception-response.decorator';
 import { GetHistoryIndicatorDto } from './dto/get-history-indicator.dto';
 import { CursorPageDto } from '../../../utils/pagination/cursor-page.dto';
-import { HistoryIndicatorDto } from '../../application/query/history-indicator/dto/history-indicator.dto';
+import { HistoryIndicatorDto } from '../../application/query/history-indicator/get-history-indicator/dto/history-indicator.dto';
 import { GetHistoryIndicatorQuery } from '../../application/query/history-indicator/get-history-indicator/get-history-indicator.query';
-import { IndicatorValueSwaggerSchema } from '../../application/query/live-indicator/dto/base-live-indicator.dto';
+import { IndicatorValueSwaggerSchema } from '../../application/query/live-indicator/get-live-indicator/dto/base-live-indicator.dto';
 
 @ApiTags('HistoryIndicatorController')
 @Controller('/api/numerical-guidance')
