+++ api/src/numerical-guidance/application/port/external/twelve/load-live-indicator.port.ts
