+++ api/src/numerical-guidance/application/port/persistence/indicator/load-history-indicator.port.ts
@@ -1,6 +1,6 @@
 import { Interval } from '../../../../../utils/type/type-definition';
 import { CursorPageDto } from '../../../../../utils/pagination/cursor-page.dto';
-import { HistoryIndicatorDto } from '../../../query/history-indicator/dto/history-indicator.dto';
+import { HistoryIndicatorDto } from '../../../query/history-indicator/get-history-indicator/dto/history-indicator.dto';
 
 export interface LoadHistoryIndicatorPort {
   loadHistoryIndicator(
