+++ api/src/numerical-guidance/application/query/history-indicator/get-history-indicator/dto/history-indicator.dto.ts
@@ -1,5 +1,5 @@
-import { Indicator } from '../../indicator/basic/dto/indicator.dto';
-import { IndicatorValue } from '../../../../../utils/type/type-definition';
+import { IndicatorValue } from '../../../../../../utils/type/type-definition';
+import { Indicator } from '../../../indicator/basic/dto/indicator.dto';
 
 export type HistoryIndicatorValue = {
   date: Date;
