--- api/src/numerical-guidance/application/query/indicator/basic/get-indicator/get-indicators.query.ts
@@ -1,5 +1,0 @@
-import { IQuery } from '@nestjs/cqrs';
-
-export class GetIndicatorsQuery implements IQuery {
-  constructor() {}
-}
