+++ api/src/numerical-guidance/application/query/live-indicator/get-live-indicator/dto/base-live-indicator.dto.ts
