+++ api/src/numerical-guidance/application/query/live-indicator/get-live-indicator/dto/live-etf.dto.ts
