+++ api/src/numerical-guidance/application/query/live-indicator/get-live-indicator/dto/live-fund.dto.ts
