+++ api/src/numerical-guidance/application/query/live-indicator/get-live-indicator/dto/live-indices.dto.ts
