+++ api/src/numerical-guidance/application/query/live-indicator/get-live-indicator/dto/live-stock.dto.ts
