+++ api/src/numerical-guidance/infrastructure/adapter/persistence/history-indicator/history-indicator.persistent.adapter.ts
@@ -5,7 +5,7 @@ import { HistoryIndicatorEntity } from './entity/history-indicator.entity';
 import { Between, LessThan, LessThanOrEqual, Repository } from 'typeorm';
 import { HistoryIndicatorValueEntity } from '../history-indicator-value/entity/history-indicator-value.entity';
 import { CursorPageDto } from '../../../../../utils/pagination/cursor-page.dto';
-import { HistoryIndicatorDto } from '../../../../application/query/history-indicator/dto/history-indicator.dto';
+import { HistoryIndicatorDto } from '../../../../application/query/history-indicator/get-history-indicator/dto/history-indicator.dto';
 import { CursorPageMetaDto } from '../../../../../utils/pagination/cursor-page.meta.dto';
 import { IndicatorValue, Interval } from '../../../../../utils/type/type-definition';
 import { HistoryIndicatorMapper } from './mapper/history-indicator.mapper';
