+++ api/src/numerical-guidance/test/integration-test/api/numerical-guidance.controller.spec.ts
@@ -2,7 +2,7 @@ import { Test } from '@nestjs/testing';
 import { CqrsModule } from '@nestjs/cqrs';
 import * as request from 'supertest';
 import { HttpStatus, INestApplication, ValidationPipe } from '@nestjs/common';
-import { LiveStockDto } from '../../../application/query/live-indicator/dto/live-stock.dto';
+import { LiveStockDto } from '../../../application/query/live-indicator/get-live-indicator/dto/live-stock.dto';
 import { liveIndicatorTestData } from '../../data/liveIndicator.test.data';
 import { CreateIndicatorBoardMetadataCommandHandler } from '../../../application/command/indicator-board-metadata/create-indicator-board-metadata/create-indicator-board-metadata.command.handler';
 import { InsertIndicatorIdCommandHandler } from '../../../application/command/indicator/insert-indicator-id/insert-indicator-id.command.handler';
